requirements.txt
pandas==2.2.2
requests==2.32.3
pydantic==2.7.4
python-dotenv==1.0.1
schedule==1.2.2